Make network share default a little safer
Set network share by default to something that will probably always be
an invalid UNC path -- so that at least we will never overwrite
something or put a file where it shouldn't be.

Also hopefully this is a little more transparent to a script user who
doesn't really know what they're doing, versus having a relatively
cryptic "." as the default path.

Better would be to actually check the validity of the network share
path before we write anything, and fail with an error message if it's
invalid.